Feature/validate

+++ Service/Transformer/TemplateTransformer.php
@@ -3,22 +3,29 @@
 namespace TemplateMakerBundle\Service\Transformer;
 
 use Pimcore\Model\AbstractModel;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+use TemplateMakerBundle\Exception\Validator\ElementValidationException;
+use TemplateMakerBundle\Exception\Validator\TemplateValidationException;
 use TemplateMakerBundle\Model\DataObject\Template;
 
 class TemplateTransformer implements TransformerInterface
 {
     private AbstractModel $model;
+
     /**
      * @param ElementTransformer $elementTransformer
+     * @param ValidatorInterface $validator
      */
     public function __construct(
-        private ElementTransformer $elementTransformer
+        private ElementTransformer $elementTransformer,
+        private ValidatorInterface $validator
     ){}
 
 
     /**
      * @param array $data
      * @return void
+     * @throws TemplateValidationException
      */
     public function transform(array $data): void
     {
@@ -27,11 +34,15 @@ class TemplateTransformer implements TransformerInterface
             ->setValue('name',$data['name'])
             ->setValue('class', $data['class']);
 
+        // validation de l'objet
+        $errors = $this->validator->validate($this->model);
+        if (count($errors)>0) {
+            $errorsString = (string) $errors;
+            throw new TemplateValidationException($errorsString);
+        }
+
         $this->model->save();
 
-        /**
-         * @todo setter le nouvel id du template sur les elements
-         */
         if (!empty($elements = $data["elements"])) {
             $this->model->getDao()->removeElements($this->model->getId());
             $this->model->setValue('elements',$this->transformElements($elements));
@@ -43,6 +54,7 @@ class TemplateTransformer implements TransformerInterface
     /**
      * @param array $elements
      * @return array
+     * @throws ElementValidationException
      */
     private function transformElements(array $elements) : array {
         $return = [];
